Material Graph for Forcefield base effects
For the base translucent effect I added a depth fade, 1-x and add to the opacity. The fog effect can be manually changed using the FadeDistance and opacity.
For the general Emissive Color I added a color and lerp (B) into the emissive color node.
For the texture animation around the shield, I added a placeholder texture "Noise Texture" from the unreal texture folder, then I connected a Texture Coord and Noise Tiling into the (A,B) of Multiply, then into the Panner, then the UVs Texture sample. The Texture sample R connects to the alpha of the Lerp (Emissive Color) and B of the Depth fade Add node. (Opacity)