Assigned color to the ForceField Textures
Adding a Param and assigning it as the Color Curve (Connecting to the Noise Texture and (A) lerp node.
I created a Color Curve and Curve Atlas to Assign to the Color curve Node.
In the Color Curve I added a red and black gradient and assigned it to the Color Curve.