Overlay Texture for Disappearing and reappearing texture
Using the Global Scaler Parameter in the M_ForceField_Instance, I activated the Alpha and adjusted the Opacity Power (Power added to its node) and Opacity Tiling ( Tiling Parameter was added to the Texture). Now there is a gradient of the Forcefield appearing and disappearing (By scaling the Alpha in the ForceField Instance). This can be applied as an animation on Player input ( To activate and deactivate the forcefield)