ForceField Blueprint and Fixed ThirdPerson to First Person
Created a ForceField Actor Blueprint and added a Timeline from 0-0 and another point in the track for 1-5. Then I added a Sphere Material (With the ForceField Instance added to it) and attatched it to the Timeline. Then created 2 custom events Show ForceField and Hide ForceField and added them to the Play and Reverse Animation Alpha (Animation Track). On update of the Anime Alpha I've attatched it to a Set Scaler Parameter Value on Materials. ( Both alphas are also conntected to each other). On finished on the Anime Alpha, A call function Hide ForceField is activated and the target is the Sphere. I also changed the Third person Boom camera to a First person camera.