Third Person to First Person Issues
Due to not checking the level status when starting the project, I picked the Third person level instead of the First person level therefore I have no weapon to test the forcefield with. I will need to create a new project in First person and also a new Repository to continue my progress.